refactor: Move storage item loans to own file
